requirements.txt
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
python-dateutil==2.8.2
pytz==2023.3